refactor: move data layers interfaces to folders
