+++ src/data/usecases/add-account/db-add-account-protocols.ts
@@ -1,4 +1,4 @@
-export * from "@/data/protocols/add-account-repository";
-export * from "@/data/protocols/encrypter";
+export * from "@/data/protocols/criptography/encrypter";
+export * from "@/data/protocols/db/add-account-repository";
 export * from "@/domain/models/account";
 export * from "@/domain/usecases/add-account";
