+++ src/infra/criptography/bcrypt-adapter.ts
@@ -1,4 +1,4 @@
-import { Encrypter } from "@/data/protocols/encrypter";
+import { Encrypter } from "@/data/protocols/criptography/encrypter";
 import bcrypt from "bcrypt";
 
 export class BcryptAdapter implements Encrypter {
