+++ src/infra/db/mongodb/account-repository/account.ts
@@ -1,4 +1,4 @@
-import { AddAccountRepository } from "@/data/protocols/add-account-repository";
+import { AddAccountRepository } from "@/data/protocols/db/add-account-repository";
 import { AccountModel } from "@/domain/models/account";
 import { AddAccountModel } from "@/domain/usecases/add-account";
 import { MongoHelper } from "../helpers/mongo-helper";
